refactor: update import paths to reflect new module structure for EN 1993-1-1:2022
