+++ blueprints/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/formula_8_3.py
@@ -3,7 +3,7 @@
 import operator
 from collections.abc import Callable
 
-from blueprints.codes.eurocode.en_1993_1_1_2025 import EN_1993_1_1_2022
+from blueprints.codes.eurocode.en_1993_1_1_2022 import EN_1993_1_1_2022
 from blueprints.codes.formula import ComparisonFormula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import KN, KNM
